# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /questions/{slug}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get question detail by slug (increments view count)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Question detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
  /answers/{id}/vote:
    post:
      operationId: voteAnswer
      tags: [answers]
      summary: Vote on an answer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "200":
          description: New score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
  /answers/{id}/accept:
    post:
      operationId: acceptAnswer
      tags: [answers]
      summary: Mark answer as accepted (question author only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Accepted answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
        "403":
          description: Not the question author
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /comments:
    post:
      operationId: createComment
      tags: [comments]
      summary: Comment on a question or answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message: { type: string }
      required: [message]
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        reputation: { type: integer }
        role:
          type: string
          enum: [member, moderator, admin, platform_owner]
        plan:
          type: string
          enum: [free, pro]
        isSuspended: { type: boolean }
        questionCount: { type: integer }
        answerCount: { type: integer }
        acceptedAnswerCount: { type: integer }
        createdAt: { type: string }
      required:
        [id, username, displayName, reputation, role, plan, isSuspended, questionCount, answerCount, acceptedAnswerCount, createdAt]
...
    UserProfileUpdate:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 80 }
        bio: { type: string, maxLength: 1000 }
...
    Question:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        body: { type: string }
        language: { type: string }
        author:
          $ref: "#/components/schemas/User"
        tags:
          type: array
          items: { type: string }
        category:
          oneOf:
            - $ref: "#/components/schemas/CategoryRef"
            - type: "null"
        score: { type: integer }
        answerCount: { type: integer }
        viewCount: { type: integer }
        hasAcceptedAnswer: { type: boolean }
        isFeatured: { type: boolean }
        myVote: { type: ["integer", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        [id, slug, title, body, language, author, tags, score, answerCount, viewCount, hasAcceptedAnswer, isFeatured, createdAt, updatedAt]
    QuestionDetail:
      type: object
      properties:
        question:
          $ref: "#/components/schemas/Question"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      required: [question, answers, comments]
...
    Answer:
      type: object
      properties:
        id: { type: integer }
        questionId: { type: integer }
        body: { type: string }
        language: { type: string }
        author:
          $ref: "#/components/schemas/User"
        score: { type: integer }
        isAccepted: { type: boolean }
        myVote: { type: ["integer", "null"] }
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt: { type: string }
      required: [id, questionId, body, language, author, score, isAccepted, comments, createdAt]
...
    Comment:
      type: object
      properties:
        id: { type: integer }
        parentType:
          type: string
          enum: [question, answer]
        parentId: { type: integer }
        body: { type: string }
        author:
          $ref: "#/components/schemas/User"
        createdAt: { type: string }
      required: [id, parentType, parentId, body, author, createdAt]
    CommentInput:
      type: object
      properties:
        parentType:
          type: string
          enum: [question, answer]
        parentId: { type: integer }
        body: { type: string, minLength: 5, maxLength: 1000 }
        website: { type: string, description: "Honeypot field - must be empty" }
      required: [parentType, parentId, body]
    VoteInput:
      type: object
      properties:
        value:
          type: integer
          enum: [-1, 0, 1]
      required: [value]
    VoteResult:
      type: object
      properties:
        score: { type: integer }
        myVote: { type: integer }
      required: [score, myVote]
...
    CategoryRef:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
      required: [slug, name]
...